+++ proyectos_de_ley/stats/management/commands/create_stats.py
@@ -54,6 +54,9 @@ class Command(BaseCommand):
         # Get projects that did not go to 2da round of votes
         self.get_dispensed_projects()
 
+        # Proyectos con dictamen pero sin votación
+        self.get_with_dictamen_but_not_voted()
+
     def get_dispensed_projects(self):
         total_approved = Seguimientos.objects.filter(evento__icontains='aprobado').count()
         total_dispensed = Seguimientos.objects.filter(evento__icontains='dispensado 2da').count()
@@ -72,3 +75,39 @@ class Command(BaseCommand):
                 'dispensed_others': dispensed_others,
             }
         )
+
+    def get_with_dictamen_but_not_voted(self):
+        """
+        Crea tabla con lista de: proyectos que no figure
+          "publicado" || promulgado || votación || en lista de seguimientos,
+        pero tenga "dictamen".
+        """
+        queryset = Seguimientos.objects.all().order_by('proyecto_id').values('proyecto_id', 'evento')
+        proyect_ids = self.get_proyect_ids(queryset)
+
+        print(self.has_dictamen(2105, queryset))
+        print(self.is_voted(2105, queryset))
+
+    def get_proyect_ids(self, queryset):
+        proyect_ids = set()
+        for i in queryset:
+            proyect_ids.add(i['proyecto_id'])
+        return proyect_ids
+
+    def is_voted(self, proyect_id, queryset):
+        events = []
+        for i in queryset:
+            if i['proyecto_id'] == proyect_id:
+                events.append(i['evento'])
+        for i in events:
+            if 'publicado' in i.lower() or 'promulgado' in i.lower() or 'votación' in i.lower():
+                return True
+
+    def has_dictamen(self, proyect_id, queryset):
+        events = []
+        for i in queryset:
+            if i['proyecto_id'] == proyect_id:
+                events.append(i['evento'])
+        for i in events:
+            if 'dictamen' in i.lower():
+                return True